+++ packages/design-system/i18next-scanner.config.js
@@ -2,12 +2,13 @@
 const typescriptTransform = require('i18next-scanner-typescript');
 
 module.exports = {
-	input: ['src/**/*.{ts,tsx}'],
+	input: ['src/**/*.{ts,tsx}', '!src/**/*stories.{ts,tsx}'],
 	options: {
+		compatibilityJSON: 'v4',
 		debug: true,
 		func: {
 			list: ['t', 'i18next.t'],
-			extensions: ['.js'],
+			extensions: ['.js', '.ts', '.tsx'],
 		},
 		lngs: ['en'],
 		defaultNs: 'design-system',