+++ packages/cmf-cqrs/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @talend/react-cmf-cqrs
 
+## 10.0.0
+
+### Major Changes
+
+- 96d688489: React: Upgrade to react 18 and @types/react 18
+
+### Patch Changes
+
+- Updated dependencies [96d688489]
+  - @talend/react-cmf@8.0.0
+
 ## 9.0.2
 
 ### Patch Changes